requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.1.2.30
Pillow==8.4.0
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
